working on float conversion

+++ o5e/headers/Base_Values_OPS.h
@@ -5,9 +5,9 @@
 #define Gasoline_SG   5642             // cc/g bin 12
 #define gram_STP_Air_Per_cc 21642 //gram STP air/cc - this is a bin24
 
-extern uint32_t Base_Pulse_Width;
-extern uint32_t Inverse_Injector_Pressure; //bin 20
-extern uint32_t Injector_Flow;
+extern float Base_Pulse_Width;
+extern float Inverse_Injector_Pressure; //bin 20
+extern float Injector_Flow;
 
 void Get_Base_Pulse_Width(void);
